# backend/docs/swagger.yaml
basePath: /
definitions:
...
  room.AppointAdminRequest:
    properties:
      newAdminId:
        type: string
    type: object
  room.CreateRoomRequest:
    properties:
      icon:
        type: string
      name:
        type: string
    type: object
  room.ParticipantResponse:
    properties:
      avatar:
        type: string
      code:
        type: string
      id:
        type: string
      username:
        type: string
    type: object
  room.RoomResponse:
    properties:
      adminId:
        type: string
      createdAt:
        type: string
      icon:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  room.SendRoomInviteRequest:
    properties:
      code:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: alisfrk2004@gmail.com
    name: Дарья
  description: Серверная часть веб-приложения для коммуникации.
  termsOfService: http://swagger.io/terms/
  title: Buzz API
  version: "1.0"
paths:
...
  /rooms:
    get:
      description: Возвращает все комнаты, в которых участвует текущий пользователь.
        Для приватных комнат название и иконка заменяются на данные собеседника.
      produces:
      - application/json
      responses:
        "200":
          description: Список комнат
          schema:
            items:
              $ref: '#/definitions/room.RoomResponse'
            type: array
        "401":
          description: Неавторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Получить список комнат пользователя
      tags:
      - rooms
  /rooms/{id}:
    delete:
      description: Полностью удаляет комнату. Доступно только администратору комнаты.
      parameters:
      - description: ID комнаты
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Комната удалена
        "400":
          description: Неверный ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Неавторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Пользователь не администратор
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Комната не найдена
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Удалить комнату
      tags:
      - rooms
    get:
      description: Возвращает детали комнаты по ID (доступно любому пользователю,
        даже не участнику, но возвращает базовую информацию).
      parameters:
      - description: ID комнаты
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Информация о комнате
          schema:
            $ref: '#/definitions/room.RoomResponse'
        "400":
          description: Неверный ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Неавторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Комната не найдена
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Получить информацию о комнате
      tags:
      - rooms
  /rooms/{id}/admin:
    post:
      consumes:
      - application/json
      description: Передаёт права администратора другому участнику комнаты. Доступно
        только текущему администратору.
      parameters:
      - description: ID комнаты
        in: path
        name: id
        required: true
        type: string
      - description: ID нового администратора
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/room.AppointAdminRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Администратор назначен
        "400":
          description: Неверные данные
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Неавторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Пользователь не администратор
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Комната не найдена
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Назначить нового администратора комнаты
      tags:
      - rooms
...
  /rooms/{id}/join:
    post:
      description: Позволяет пользователю немедленно присоединиться к комнате (без
        ожидания подтверждения). Комната должна существовать.
      parameters:
      - description: ID комнаты
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Успешное вступление
        "400":
          description: Неверный ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Неавторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Комната не найдена
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Пользователь уже участник
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Вступить в комнату по ссылке-приглашению
      tags:
      - rooms
  /rooms/{id}/leave:
    post:
      description: Выход текущего пользователя из комнаты. Если комната приватная,
        она будет удалена, а дружба разорвана. Если администратор покидает публичную
        комнату, права передаются другому участнику.
      parameters:
      - description: ID комнаты
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Выход выполнен
        "400":
          description: Неверный ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Неавторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Пользователь не является участником
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Комната не найдена
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Покинуть комнату
      tags:
      - rooms
  /rooms/{id}/participants:
    get:
      description: Возвращает всех участников указанной комнаты.
      parameters:
      - description: ID комнаты
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Список участников
          schema:
            items:
              $ref: '#/definitions/room.ParticipantResponse'
            type: array
        "400":
          description: Неверный ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Неавторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Получить список участников комнаты
      tags:
      - rooms
  /rooms/{id}/participants/{userId}:
    delete:
      description: Удаляет указанного пользователя из комнаты. Доступно только администратору
        комнаты.
      parameters:
      - description: ID комнаты
        in: path
        name: id
        required: true
        type: string
      - description: ID пользователя для удаления
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Участник удалён
        "400":
          description: Неверные параметры или попытка удалить себя
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Неавторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Пользователь не администратор
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Комната не найдена
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Удалить участника из комнаты
      tags:
      - rooms
  /rooms/{id}/send-invite:
    post:
      consumes:
      - application/json
      description: Отправляет запрос на вступление в комнату другому пользователю.
        Приглашающий должен быть участником комнаты.
      parameters:
      - description: ID комнаты
        in: path
        name: id
        required: true
        type: string
      - description: Данные приглашаемого
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/room.SendRoomInviteRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Приглашение отправлено
        "400":
          description: Некорректные данные или попытка пригласить себя
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Неавторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Пользователь не является участником комнаты
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Комната или пользователь не найдены
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Пользователь уже участник или приглашение уже существует
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Отправить приглашение в комнату
      tags:
      - rooms
...
  /rooms/create:
    post:
      consumes:
      - application/json
      description: Создаёт комнату с указанным названием (и опциональной иконкой).
        Текущий пользователь становится администратором и первым участником.
      parameters:
      - description: Данные комнаты
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/room.CreateRoomRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Комната создана
        "400":
          description: Некорректные данные
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Неавторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Создать новую комнату
      tags:
      - rooms
...
securityDefinitions:
  BearerAuth:
    description: Введите токен в формате "Bearer <token>"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"